+++ changelog.md
@@ -23,11 +23,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 - Nothing.
 
-## [0.0.0] - 2019-01-01
-- Nothing.
+## [1.0.0-preview] - 2019-02-17
+### Added
+- Migrate from internal repository.
+- Add non-alloc GetEnumerator. (#3)
+
+### Changed
+- Make InstanceCollection as read-only. (#5)
+- Split default collections to non-cached and cached. (#4)
+- Initialize IdentifierType on creation. (#7)
+
+### Removed
+- Remove InstanceIdentifierGeneratorPrimitive. (#6)
 
 ---
 > Unity Game Framework | Copyright 2019
 
-[Unreleased]: https://github.com/unity-game-framework/ugf-draft/compare/0.0.0...HEAD
-[0.0.0]: https://github.com/unity-game-framework/ugf-draft/compare/0.0.0...HEAD
+[Unreleased]: https://github.com/unity-game-framework/ugf-instance/compare/1.0.0...HEAD
+[1.0.0-preview]: https://github.com/unity-game-framework/ugf-instance/compare/master...release/1.0.0-preview